powershell basics (#92)
* start work on making pwsh functional

* remove bad test

* bad file

* more pwsh cleanup / commands

* files

+++ src/PdfLexer.Powershell/OutPdf.cs
@@ -1,5 +1,6 @@
 ﻿using PdfLexer.DOM;
 using PdfLexer.Serializers;
+using System.IO;
 using System.Management.Automation;
 
 namespace PdfLexer.Powershell;
@@ -7,11 +8,14 @@ namespace PdfLexer.Powershell;
 [Cmdlet(
         "Out",
         "Pdf"),
+    OutputType(typeof(string)
+    ),
    ]
 public class OutPdf : InputOutputPathCmdlet, IDisposable
 {
     private FileStream? _fo;
     private StreamingWriter? _sw;
+    private string? path;
 
     [Parameter(
         Mandatory = false,
@@ -25,9 +29,19 @@ public class OutPdf : InputOutputPathCmdlet, IDisposable
     [ValidateNotNullOrEmpty]
     public SwitchParameter Append { get; set; }
 
+    [Parameter(
+    Mandatory = false,
+    HelpMessage = "If documents passed should be closed / disposed after pages written.")]
+    public SwitchParameter Close { get; set; }
+
+    [Parameter(
+    Mandatory = false,
+    HelpMessage = "If file at output location should be replaced. Ignored if -Append used.")]
+    public SwitchParameter Force { get; set; }
+
     protected override void BeginProcessing()
     {
-        var path = GetOutputPath();
+        path = GetOutputPath();
         if (path == null)
         {
             return;
@@ -38,14 +52,21 @@ public class OutPdf : InputOutputPathCmdlet, IDisposable
             File.Move(path, tmp);
             _fo = File.Create(path);
             _sw = new StreamingWriter(_fo, DedupObjects, true, true);
-            using var pv = PsHelpers.OpenDocument(tmp);
-            foreach (var pg in pv.Pages)
             {
-                _sw.AddPage(pg);
+                using var pv = PdfDocument.Open(tmp);
+                foreach (var pg in pv.Pages)
+                {
+                    _sw.AddPage(pg);
+                }
             }
+            File.Delete(tmp);
         }
         else
         {
+            if (File.Exists(path) && !Force)
+            {
+                throw new PdfLexerException($"File exists at {path} and -Force option not used");
+            }
             _fo = File.Create(path);
             _sw = new StreamingWriter(_fo, DedupObjects, true, true);
         }
@@ -57,6 +78,13 @@ public class OutPdf : InputOutputPathCmdlet, IDisposable
         {
             _sw.AddPage(pg);
         }
+        if (Document != null && Close)
+        {
+            foreach (var doc in Document) 
+            {
+                doc.Dispose();
+            }
+        }
         GC.Collect();
     }
 
@@ -67,13 +95,14 @@ public class OutPdf : InputOutputPathCmdlet, IDisposable
 
     protected override void EndProcessing()
     {
-
         _sw?.Complete(new PdfDictionary());
         _sw?.Dispose();
         _fo?.Dispose();
         _sw = null;
         _fo = null;
         GC.Collect(GC.MaxGeneration, GCCollectionMode.Aggressive, true, true);
+
+        WriteObject(path);
     }
 
     public void Dispose()
